apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "restaurant-menu-api"
version = "0.1.0"
description = "FastAPI backend for the restaurant digital menu SaaS"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "asyncpg>=0.29.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.1.1",
  "fastapi>=0.115.0",
  "greenlet>=3.1.1",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.4.0",
  "psycopg[binary]>=3.2.3",
  "PyJWT>=2.9.0",
  "python-multipart>=0.0.9",
  "SQLAlchemy>=2.0.36",
  "stripe>=11.1.0",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
